+++ R/sim_monthly.R
@@ -3,8 +3,10 @@
 #' Simulate a monthly seasonal series
 #' @param N Length in years
 #' @param sd Standard deviation for all seasonal factors
+#' @param change_sd Standard deviation of shock to seasonal factor
 #' @param beta_1 Persistance wrt to previous period of the seasonal change 
-#' @param change_sd Standard deviation of simulated change for all seasonal factors
+#' @param beta_tau Persistence wrt to one year/cycle of the seasonal change
+#' @param moving Is the seasonal pattern allowed to change over time
 #' @param extra_smooth Boolean. Should the seasonal factors be smooth on a period-by-period basis
 #' @param start Start date of output time series
 #' @param multiplicative Boolean. Should multiplicative seasonal factors be simulated
@@ -23,13 +25,13 @@
 #' @references Ollech, D. (2021). Seasonal adjustment of daily time series. Journal of Time Series Econometrics. \doi{10.1515/jtse-2020-0028}
 #' @export
 
-sim_monthly <- function(N, sd=1, beta_1=0.9, change_sd=0.025,model=list(order=c(3,1,1), ma=0.5, ar=c(0.2, -0.4, 0.1)), start=c(2010,1), multiplicative=TRUE, extra_smooth=FALSE) {
+sim_monthly <- function(N, sd=5, change_sd = sd / 10, beta_1= 0.6, beta_tau = 0.4, moving = TRUE, model=list(order=c(3,1,1), ma=0.5, ar=c(0.2, -0.4, 0.1)), start=c(2010,1), multiplicative=TRUE, extra_smooth=FALSE) {
   # Basic series
   series <- stats::ts(stats::arima.sim(n=12*N-1, model=model), start=start, frequency=12)
   series <- series - min(series) + 100
   
   # Add seasonal effect
-  sfac <- sim_sfac(length(series), freq=12, sd=sd, change_sd=change_sd, beta_1= beta_1, ar=0.99, ma=0.99, start=start, burnin=3, multiplicative=multiplicative, extra_smooth=extra_smooth)
+  sfac <- sim_sfac(length(series), freq=12, sd=sd, change_sd = change_sd, moving = moving, beta_1= beta_1, beta_tau = beta_tau, ar=0.99, ma=0.99, start=start, burnin=3, multiplicative=multiplicative, extra_smooth=extra_smooth)
   
   series <- tsbox::ts_xts(series)
   seas_adj <- series