+++ NEWS.md
@@ -1,3 +1,17 @@
+tssim 0.2.0 to 0.2.7 (Release date: 2024-12-13)
+============================================
+Changes: - Added hs and mstl based simulation functions
+         - Improved arguments and default setting for sim_daily
+
+tssim 0.1.9 (Release date: 2024-05-10)
+============================================
+Changes: - Reintroduction and improvement of the change_sd parameter. It now more directly influences how quickly the seasonal component changes
+
+tssim 0.1.8 (Release date: 2024-04-30)
+============================================
+Changes: - Improved the default setting of sim_daily and sim_monthly
+	 - Restructured parts of the code of sim_sfac
+
 tssim 0.1.6 (Release date: 2021-06-21)
 ============================================
 Changes: - Updated documentation